+++ NEWS.md
@@ -21,6 +21,7 @@ New Features
 
   - `"rhombille"` implements a rhombille tiling of rhombi (#37)
   - `"tetrakis_square"` implements a tetrakis square tiling of isosceles right triangles (#38)
+  - `"2*53.2**233.2*53.2**233"` implements a polygon tiling made of rhombi.
   - `"3.3*30.3.3*30"` implements a regular (star) polygon tiling made of triangles and three-pointed stars.
   - `"3.3.3.12*30.3.3.12*30"` implements a regular (star) polygon tiling made of triangles and twelve-pointed starts.
   - `"3.3.8*15.3.4.3.8*15"` implements a regular (star) polygon tiling of triangles,
@@ -32,7 +33,7 @@ New Features
   - `"3.4.8.3.8*15"` implements a regular (star) polygon tiling of triangles,
     squares, octagons, and eight-pointed stars.
   - `"3.6*30.6**30"` implements a regular (star) polygon tiling made of triangles and six-pointed stars.
-  - `"4.2*60.4.2**240"` implements a regular (star) polygon tiling made of squares and rhombi (two-pointed stars).
+  - `"4.2*60.4.2**240"` implements a polygon tiling made of squares and rhombi.
   - `"4.4*30.4**30"` implements a regular (star) polygon tiling made of squares and four-pointed stars.
   - `"4.6.4*30.6"` implements a regular (star) polygon tiling made of squares, hexagons, and four-pointed stars.
   - `"4.6*30.4.6*30.4.6*30"` implements a regular (star) polygon of squares and six-pointed stars.